+++ Helper/ReminderHelper.cs
@@ -4,8 +4,8 @@ namespace CarCareTracker.Helper
 {
     public interface IReminderHelper
     {
-        ReminderRecord GetUpdatedRecurringReminderRecord(ReminderRecord existingReminder, DateTime? currentDate, int? currentMileage);
-        List<ReminderRecordViewModel> GetReminderRecordViewModels(List<ReminderRecord> reminders, int currentMileage, DateTime dateCompare);
+        ReminderRecord GetUpdatedRecurringReminderRecord(ReminderRecord existingReminder, DateTime? currentDate, decimal? currentMileage);
+        List<ReminderRecordViewModel> GetReminderRecordViewModels(List<ReminderRecord> reminders, decimal currentMileage, DateTime dateCompare);
     }
     public class ReminderHelper: IReminderHelper
     {
@@ -14,7 +14,7 @@ namespace CarCareTracker.Helper
         {
             _config = config;
         }
-        public ReminderRecord GetUpdatedRecurringReminderRecord(ReminderRecord existingReminder, DateTime? currentDate, int? currentMileage)
+        public ReminderRecord GetUpdatedRecurringReminderRecord(ReminderRecord existingReminder, DateTime? currentDate, decimal? currentMileage)
         {
             var newDate = currentDate ?? existingReminder.Date;
             var newMileage = currentMileage ?? existingReminder.Mileage;
@@ -74,7 +74,7 @@ namespace CarCareTracker.Helper
             }
             return existingReminder;
         }
-        public List<ReminderRecordViewModel> GetReminderRecordViewModels(List<ReminderRecord> reminders, int currentMileage, DateTime dateCompare)
+        public List<ReminderRecordViewModel> GetReminderRecordViewModels(List<ReminderRecord> reminders, decimal currentMileage, DateTime dateCompare)
         {
             List<ReminderRecordViewModel> reminderViewModels = new List<ReminderRecordViewModel>();
             var reminderUrgencyConfig = _config.GetReminderUrgencyConfig();
