+++ Models/API/VehicleInfo.cs
@@ -18,7 +18,7 @@
         public decimal TaxRecordCost { get; set; }
         public int GasRecordCount { get; set; }
         public decimal GasRecordCost { get; set; }
-        public int LastReportedOdometer { get; set; }
+        public decimal LastReportedOdometer { get; set; }
         public int PlanRecordBackLogCount { get; set; }
         public int PlanRecordInProgressCount { get; set; }
         public int PlanRecordTestingCount { get; set; }
