+++ Models/Report/CostForVehicleByMonth.cs
@@ -6,7 +6,7 @@
         public int MonthId { get; set; }
         public string MonthName { get; set; }
         public decimal Cost { get; set; }
-        public int DistanceTraveled { get; set; }
+        public decimal DistanceTraveled { get; set; }
         public decimal CostPerDistanceTraveled { get { if (DistanceTraveled > 0) { return Cost / DistanceTraveled; } else { return 0M; } } }
     }
 }
